Add warning for 0 call interval
If you set the call interval to 0,
(by accidentally doing integer division for example)
the pyglet window becomes unclosable and will hang
until you kill the process. This change adds a warning
to let users know this has happened and helps debug
what should be a simple issue but can be hard to
track down.